rename Bc4/BC5 storeproducts to GoogleStoreProduct and move into flavor folders
