+++ feature/google/src/test/java/com/revenuecat/purchases/google/StoreProductTest.kt
@@ -2,7 +2,7 @@ package com.revenuecat.purchases.google
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.revenuecat.purchases.ProductType
-import com.revenuecat.purchases.google.BC4StoreProduct
+import com.revenuecat.purchases.google.GoogleStoreProduct
 import org.assertj.core.api.Assertions
 import org.json.JSONObject
 import org.junit.Test
@@ -13,7 +13,7 @@ class StoreProductTest {
 
     @Test
     fun `Two StoreProducts with the same properties are equal`() {
-        val storeProduct1 = BC4StoreProduct(
+        val storeProduct1 = GoogleStoreProduct(
             sku = "sku",
             type = ProductType.INAPP,
             price = "$2",
@@ -32,7 +32,7 @@ class StoreProductTest {
             iconUrl = "http://",
             originalJson = JSONObject("{}")
         )
-        val storeProduct2 = BC4StoreProduct(
+        val storeProduct2 = GoogleStoreProduct(
             sku = "sku",
             type = ProductType.INAPP,
             price = "$2",
@@ -56,7 +56,7 @@ class StoreProductTest {
 
     @Test
     fun `Two StoreProducts with the same properties have the same hashcode`() {
-        val storeProduct1 = BC4StoreProduct(
+        val storeProduct1 = GoogleStoreProduct(
             sku = "sku",
             type = ProductType.INAPP,
             price = "$2",
@@ -75,7 +75,7 @@ class StoreProductTest {
             iconUrl = "http://",
             originalJson = JSONObject("{}")
         )
-        val storeProduct2 = BC4StoreProduct(
+        val storeProduct2 = GoogleStoreProduct(
             sku = "sku",
             type = ProductType.INAPP,
             price = "$2",
